+++ forui/lib/src/widgets/breadcrumb.dart
@@ -215,10 +215,10 @@ class _CollapsedCrumbState extends State<_CollapsedCrumb> with SingleTickerProvi
     final style = FBreadcrumbItemData.of(context).style;
     return Padding(
       padding: style.padding,
-      child: FPopoverMenu.tappable(
+      child: FPopoverMenu.automatic(
         style: widget.popOverMenuStyle,
         popoverController: controller,
-        menuAnchor: Alignment.topLeft,
+        popoverAnchor: Alignment.topLeft,
         childAnchor: Alignment.bottomLeft,
         menu: widget.menu,
         child: FIcon(