+++ forui/lib/src/widgets/text_field/text_field.dart
@@ -17,7 +17,7 @@ import 'package:forui/src/widgets/text_field/field.dart';
 /// * https://forui.dev/docs/form/text-field for working examples.
 /// * [FTextFieldStyle] for customizing a text field's appearance.
 /// * [TextField] for more details about working with a text field.
-final class FTextField extends StatelessWidget with FFormFieldProperties<String> {
+final class FTextField extends StatelessWidget implements FFormFieldProperties<String> {
   static Widget _contextMenuBuilder(
     BuildContext context,
     EditableTextState state,