+++ samples/lib/widgets/popover_menu.dart
@@ -32,7 +32,7 @@ class _State extends StatefulSampleState<PopoverMenuPage> with SingleTickerProvi
         actions: [
           FPopoverMenu(
             popoverController: controller,
-            menuAnchor: Alignment.topRight,
+            popoverAnchor: Alignment.topRight,
             childAnchor: Alignment.bottomRight,
             menu: [
               FTileGroup(